+++ src/imio/urban/core/contents/parcel/__init__.py
@@ -1,0 +1,7 @@
+# -*- coding: utf-8 -*-
+
+from imio.urban.core.contents.parcel.content import IParcel
+from imio.urban.core.contents.parcel.content import Parcel
+
+IParcel
+Parcel
