--- src/imio/urban/core/contents/parcel/browser/configure.zcml
@@ -1,28 +1,0 @@
-<configure
-    xmlns="http://namespaces.zope.org/zope"
-    xmlns:browser="http://namespaces.zope.org/browser"
-    i18n_domain="imio.urban.core">
-
-  <browser:page
-     for="imio.urban.core.contents.parcel.interfaces.IParcel"
-     name="view"
-     class=".views.ParcelViewRedirects"
-     permission="zope2.View"
-  />
-
-  <browser:page
-     for="imio.urban.core.contents.parcel.interfaces.IParcel"
-     name="edit"
-     class=".views.ParcelEditForm"
-     permission="zope2.View"
-  />
-
-  <browser:page
-     for="imio.urban.core.contents.parcel.interfaces.IParcel"
-     name="parcelview"
-     class=".views.ParcelView"
-     permission="zope2.View"
-     template="../../browser/item.pt"
-  />
-
-</configure>
