--- src/imio/urban/core/contents/parcel/events.zcml
@@ -1,38 +1,0 @@
-<configure xmlns="http://namespaces.zope.org/zope"
-           xmlns:five="http://namespaces.zope.org/five"
-           xmlns:i18n="http://namespaces.zope.org/i18n"
-           i18n_domain="imio.urban.core">
-
-  <subscriber for="imio.urban.core.contents.parcel.interfaces.IParcel
-                   zope.lifecycleevent.interfaces.IObjectCreatedEvent"
-               handler=".events.set_authenticity" />
-
-  <subscriber for="imio.urban.core.contents.parcel.interfaces.IParcel
-                   zope.lifecycleevent.interfaces.IObjectModifiedEvent"
-              handler=".events.set_authenticity" />
-
-  <subscriber for="imio.urban.core.contents.parcel.interfaces.IParcel
-                   zope.lifecycleevent.interfaces.IObjectModifiedEvent"
-              handler=".events.update_container_parcelindex" />
-
-  <subscriber for="imio.urban.core.contents.parcel.interfaces.IParcel
-                   zope.container.interfaces.IObjectRemovedEvent"
-              handler=".events.update_container_parcelindex" />
-
-  <subscriber for="imio.urban.core.contents.parcel.interfaces.IParcel
-                   zope.lifecycleevent.interfaces.IObjectAddedEvent"
-              handler=".events.update_bound_licences_parcelindex" />
-
-  <subscriber for="imio.urban.core.contents.parcel.interfaces.IParcel
-                   zope.lifecycleevent.interfaces.IObjectModifiedEvent"
-              handler=".events.update_bound_licences_parcelindex" />
-
-  <subscriber for="imio.urban.core.contents.parcel.interfaces.IParcel
-                   zope.container.interfaces.IObjectRemovedEvent"
-              handler=".events.update_bound_licences_parcelindex" />
-
-  <subscriber for="imio.urban.core.contents.parcel.interfaces.IParcel
-                   zope.lifecycleevent.interfaces.IObjectModifiedEvent"
-              handler=".events.update_parcelling_title" />
-
-</configure>
